+++ RedPandaIDE/settingsdialog/compilersetoptionwidget.cpp
@@ -99,6 +99,7 @@ static void loadCompilerSetSettings(Settings::PCompilerSet pSet, Ui::CompilerSet
     ui->txtGDBServer->setText(pSet->debugServer());
     ui->txtResourceCompiler->setText(pSet->resourceCompiler());
     ui->txtProfiler->setText(pSet->profiler());
+    ui->txtAssembler->setText(pSet->assembler());
 
     if (pSet->execCharset() == ENCODING_AUTO_DETECT
             || pSet->execCharset() == ENCODING_SYSTEM_DEFAULT
@@ -208,6 +209,7 @@ void CompilerSetOptionWidget::saveCurrentCompilerSet()
     pSet->setDebugServer(ui->txtGDBServer->text().trimmed());
     pSet->setResourceCompiler(ui->txtResourceCompiler->text().trimmed());
     pSet->setProfiler(ui->txtProfiler->text().trimmed());
+    pSet->setAssembler(ui->txtAssembler->text().trimmed());
 
     pSet->binDirs()=mBinDirWidget->dirList();
 
@@ -334,6 +336,7 @@ void CompilerSetOptionWidget::updateIcons(const QSize& /*size*/)
     pIconsManager->setIcon(ui->btnChooseGDBServer, IconsManager::ACTION_FILE_OPEN_FOLDER);
     pIconsManager->setIcon(ui->btnChooseMake, IconsManager::ACTION_FILE_OPEN_FOLDER);
     pIconsManager->setIcon(ui->btnChooseProfiler, IconsManager::ACTION_FILE_OPEN_FOLDER);
+    pIconsManager->setIcon(ui->btnChooseAssembler, IconsManager::ACTION_FILE_OPEN_FOLDER);
     pIconsManager->setIcon(ui->btnChooseResourceCompiler, IconsManager::ACTION_FILE_OPEN_FOLDER);
 }
 
@@ -443,5 +446,17 @@ void CompilerSetOptionWidget::on_btnChooseProfiler_clicked()
                 tr("Executable files (*.exe)"));
     if (fileExists(fileName))
         ui->txtProfiler->setText(fileName);
+}
+
+
+void CompilerSetOptionWidget::on_btnChooseAssembler_clicked()
+{
+    QString fileName = QFileDialog::getOpenFileName(
+                this,
+                tr("Locate nasm"),
+                getBinDir(),
+                tr("Executable files (*.exe)"));
+    if (fileExists(fileName))
+        ui->txtAssembler->setText(fileName);
 }